Add TS dependency and deploy commands
TS dependency was added, so anyone using the examples can run them out
of the box, rather than having to add is as a global dependency.

Custom scripts in the package.json allow people to run the example
scripts as soon as they install the dependencies. This way they can
experience the expected behaviour before digging into the code.